==> pBasics/p02-KM.Rmd <==
# KM {#p02}

```{r 'P02-R', decorate=TRUE, include=FALSE, cache=FALSE}
sys.source(here::here("code", "A00Knitr.R"), envir = knitr::knit_global())
sys.source(here::here("code", "A01Packages.R"), envir = knitr::knit_global())
sys.source(here::here("code", "A02AllUDF.R"), envir = knitr::knit_global())
reticulate::source_python(here::here("code", "B01Modules.py"), convert = FALSE)
sys.source(here::here("code", "A04PackagesPython.R"), envir = knitr::knit_global())
if(FALSE) knit_exit() #Evaluate only up to this chunk and complete the knit process
```

## Kite

```{r 'P02-R-Kite', decorate=TRUE, include=FALSE}
bb <- make_graph("Krackhardt kite")
V(bb)$name <- LETTERS[1:vcount(bb)]
r_kite <- bb
```

```{python 'P02-Y-Kite', decorate=TRUE, include=FALSE}
pp = ig.Graph.Famous("Krackhardt_Kite")
pp.vs['name'] = list(string.ascii_uppercase[0:pp.vcount()])
pp.vs['label'] = pp.vs['name']
y_kite = copy.deepcopy(pp)

```

```{r 'P02-R-Kite-Print', decorate=TRUE}
print(r_kite)
```

```{python 'P02-Y-Kite-Print', decorate=TRUE}
print(y_kite)

```

## KM Network

\textcolor{pink}{Please import the "P01-KM-Week1.xlsx" and "P01-KM-Week4.xlsx"}

- About: Edges and Vertices of a Class Network at Week 1 and Week 4

```{r 'P02-R-XL-KM', decorate=TRUE, include=FALSE}
loc_rds <- here("data", "P01-R-KM-Week1.rds")
stopifnot(file.exists(loc_rds))
p02r_km_1 <- readRDS(loc_rds)
loc_rds <- here("data", "P01-R-KM-Week4.rds")
stopifnot(file.exists(loc_rds))
p02r_km_4 <- readRDS(loc_rds)
```

```{r 'P02-R-Dimensions-KM', decorate=TRUE}
# #Dimensions of Each Sheet: Row 1 (Rows), Row 2 (Columns), Columns (Sheets)
vapply(p02r_km_1, dim, numeric(2))
vapply(p02r_km_4, dim, numeric(2))
# #Structure
if(FALSE) str(p02r_km_1)
if(FALSE) str(p02r_km_4)
# #Summary of List Objects
if(FALSE) lapply(p02r_km_1, summary)
if(FALSE) lapply(p02r_km_4, summary)
```

## Eigenvector Centrality vs. Degree, Betweenness, Closeness

- NOTE: R (igraph) and Python (NetworkX) Eigenvalues match each other but not with NodeXL
  - R & Python: $\{A, B, C, D, E\} = \{0.653, 0.354, 0.354, 0.500, 0.271\}$
  - NodeXL: $\{A, B, C, D, E\} = \{0.562, 0.397, 0.397, 0.562, 0.233\}$
  
### Y Network

```{r 'P02-R-Centralities-NetY', decorate=TRUE}
bb <- graph_from_literal(E-D-A-B, A-C)
tibble(ID = 1:vcount(bb), Nodes = V(bb)$name, Degree = degree(bb),
       Closeness = round(closeness(bb, normalized = FALSE), 3),
       Betweenness = betweenness(bb),
       Eigenvalues = round(eigen_centrality(bb, scale = FALSE)$vector, 3))
```


```{python 'P02-Y-NetY', decorate=TRUE}
# #Create NetworkX
nx_g = nx.Graph()
nx_g.add_nodes_from(list(string.ascii_uppercase[0:5]))
nx_g.add_edges_from([('E', 'D'), ('D', 'A'), ('A', 'B'), ('A', 'C')])

# #Convert NetworkX to iGraph
ig_g = ig.Graph.TupleList(nx_g.edges())

print(nx_g)
print(nx_g.nodes)
print(nx_g.edges)

nn_nod = dict(zip(nx_g.nodes, nx_g.nodes, strict = True))
ii_dgr = dict(nx_g.degree())
jj = nx.closeness_centrality(nx_g)
jj_clo = {k:round(v, 3) for k, v in jj.items()}
kk = [round(i, 3) for i in ig_g.closeness(normalized = False)]
kk_clo = dict(zip(ig_g.vs['name'], kk, strict = True))
ll_bet = nx.betweenness_centrality(nx_g, normalized = False) 
mm = nx.eigenvector_centrality(nx_g)
mm_eig = {k:round(v, 3) for k, v in mm.items()}
#
oo = pd.concat( [pd.Series(x) for x in (nn_nod, ii_dgr, jj_clo, kk_clo, ll_bet, mm_eig)], axis=1)
oo.columns = ("Nodes", "Degree", "Closeness_N", "Closeness", "Betweenness", "Eigenvector")
print(oo)

```

### Kite

```{r 'P02-R-Centralities-Kite', decorate=TRUE}
bb <- r_kite
tibble(ID = 1:vcount(bb), Nodes = V(bb)$name, Degree = degree(bb),
       Closeness = round(closeness(bb, normalized = FALSE), 3),
       Betweenness = betweenness(bb),
       Eigenvalues = round(eigen_centrality(bb, scale = FALSE)$vector, 3))
```

## Validation {#val-p02 .unlisted .unnumbered .tabset .tabset-fade}

```{r 'P02-R-Cleanup', decorate=TRUE, include=FALSE, cache=FALSE}
if(FALSE) f_rmExist(aa)
```

```{r 'P02-R-Validation', decorate=TRUE, include=TRUE, cache=FALSE}
# #SUMMARISED Packages and Objects (BOOK CHECK)
f_()
#
difftime(Sys.time(), q_start)
```

****
